backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.9.2
scikit-learn>=1.4.0
pandas>=2.0.0
joblib>=1.3.0
